@@ -21,10 +21,12 @@
   "author": "Artem Lukoianov, Davit Martirosyan, Fedor Moiseev",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.9.0",
+    "@babel/core": "^7.9.6",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^5.1.1",
@@ -46,10 +48,11 @@
   },
   "dependencies": {
     "@babel/polyfill": "^7.8.7",
+    "babel-preset-react": "^6.24.1",
     "core-js": "^3.6.5",
     "d3": "^5.15.1",
     "puppeteer": "^3.0.0",
-    "react": "^16.8.0",
-    "react-dom": "^16.8.0"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
   }
 }
